+++ tests/sans2d/SANS2D00063091.SolidAngle_from_mantid.hdf5
